impl excel hyperlinks to html user-guide
