+++ pom.xml
@@ -172,7 +172,7 @@
                             <goal>process-asciidoc</goal>
                         </goals>
                         <configuration>
-                            <outputDirectory>target/classes/static</outputDirectory>
+                            <outputDirectory>target/classes/static/user-guide/</outputDirectory>
                         </configuration>
                     </execution>
                 </executions>
