+++ src/main/asciidoc/stock-market-profit.adoc
