+++ src/main/java/ru/investbook/view/excel/PortfolioPaymentExcelTableFactory.java
@@ -22,11 +22,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
-import ru.investbook.converter.EventCashFlowConverter;
+import ru.investbook.converter.SecurityEventCashFlowConverter;
+import ru.investbook.entity.SecurityEntity;
 import ru.investbook.pojo.CashFlowType;
-import ru.investbook.pojo.EventCashFlow;
 import ru.investbook.pojo.Portfolio;
-import ru.investbook.repository.EventCashFlowRepository;
+import ru.investbook.pojo.SecurityEventCashFlow;
+import ru.investbook.repository.SecurityEventCashFlowRepository;
+import ru.investbook.repository.SecurityRepository;
 import ru.investbook.view.Table;
 import ru.investbook.view.TableFactory;
 
@@ -42,37 +44,37 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static ru.investbook.pojo.CashFlowType.*;
-import static ru.investbook.view.excel.ForeignPortfolioPaymentTableHeader.*;
+import static ru.investbook.view.excel.PortfolioPaymentExcelTableHeader.*;
 
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class ForeignPortfolioPaymentTableFactory implements TableFactory {
-    /** tax accounted by {@link TaxExcelTableFactory} */
-    private static final CashFlowType[] PAY_TYPES = new CashFlowType[]{AMORTIZATION, REDEMPTION, COUPON, DIVIDEND};
-    private final EventCashFlowRepository eventCashFlowRepository;
-    private final EventCashFlowConverter eventCashFlowConverter;
+public class PortfolioPaymentExcelTableFactory implements TableFactory {
+    private static final CashFlowType[] PAY_TYPES = new CashFlowType[]{AMORTIZATION, REDEMPTION, COUPON, DIVIDEND, TAX};
+    private final SecurityEventCashFlowRepository securityEventCashFlowRepository;
+    private final SecurityEventCashFlowConverter securityEventCashFlowConverter;
+    private final SecurityRepository securityRepository;
     private final ForeignExchangeRateTableFactory foreignExchangeRateTableFactory;
 
     @Override
     public Table create(Portfolio portfolio) {
-        List<EventCashFlow> cashFlows = getCashFlows(portfolio);
+        List<SecurityEventCashFlow> cashFlows = getCashFlows(portfolio);
         return getTable(cashFlows);
     }
 
-    private ArrayList<EventCashFlow> getCashFlows(Portfolio portfolio) {
-        return eventCashFlowRepository
+    private ArrayList<SecurityEventCashFlow> getCashFlows(Portfolio portfolio) {
+        return securityEventCashFlowRepository
                 .findByPortfolioIdAndCashFlowTypeIdInOrderByTimestampDesc(
                         portfolio.getId(),
                         Stream.of(PAY_TYPES)
                                 .map(CashFlowType::getId)
                                 .collect(Collectors.toList()))
                 .stream()
-                .map(eventCashFlowConverter::fromEntity)
+                .map(securityEventCashFlowConverter::fromEntity)
                 .collect(Collectors.toCollection(ArrayList::new));
     }
 
-    private Table getTable(List<EventCashFlow> cashFlows) {
+    private Table getTable(List<SecurityEventCashFlow> cashFlows) {
         Table table = new Table();
         if (!cashFlows.isEmpty()) {
             table.add(new Table.Record());
@@ -80,7 +82,7 @@ public class ForeignPortfolioPaymentTableFactory implements TableFactory {
             table.add(monthTotalRecord);
             Month month = null;
             int sumRowCount = 0;
-            for (EventCashFlow cash : cashFlows) {
+            for (SecurityEventCashFlow cash : cashFlows) {
                 Instant timestamp = cash.getTimestamp();
                 Month currentMonth = LocalDate.ofInstant(timestamp, ZoneId.systemDefault()).getMonth();
                 if (month == null) {
@@ -95,11 +97,13 @@ public class ForeignPortfolioPaymentTableFactory implements TableFactory {
                 }
                 Table.Record record = new Table.Record();
                 record.put(DATE, timestamp);
-                record.put(ForeignPortfolioPaymentTableHeader.CASH, cash.getValue());
+                record.put(COUNT, cash.getCount());
+                record.put(PortfolioPaymentExcelTableHeader.CASH, cash.getValue());
                 record.put(CURRENCY, cash.getCurrency());
                 record.put(CASH_RUB, foreignExchangeRateTableFactory.cashConvertToRubExcelFormula(cash.getCurrency(),
-                        ForeignPortfolioPaymentTableHeader.CASH, EXCHANGE_RATE));
-                record.put(DESCRIPTION, cash.getDescription());
+                        PortfolioPaymentExcelTableHeader.CASH, EXCHANGE_RATE));
+                record.put(PAYMENT_TYPE, getPaymentType(cash));
+                record.put(SECURITY, getSecurityName(cash));
                 table.add(record);
                 sumRowCount++;
             }
@@ -113,8 +117,25 @@ public class ForeignPortfolioPaymentTableFactory implements TableFactory {
 
     private void calcTotalRecord(Table.Record monthTotalRecord, Month month, int sumRowCount) {
         if (sumRowCount != 0) {
-            monthTotalRecord.put(DATE, StringUtils.capitalize(month.getDisplayName(TextStyle.FULL_STANDALONE, Locale.getDefault())));
+            monthTotalRecord.put(SECURITY, StringUtils.capitalize(month.getDisplayName(TextStyle.FULL_STANDALONE, Locale.getDefault())));
             monthTotalRecord.put(CASH_RUB, "=SUM(OFFSET(" + CASH_RUB.getCellAddr() + ",1,0," + sumRowCount + ",1))");
         }
+    }
+
+    private static String getPaymentType(SecurityEventCashFlow cash) {
+        return switch (cash.getEventType()) {
+            case DIVIDEND -> "Дивиденды";
+            case COUPON -> "Купоны";
+            case REDEMPTION -> "Погашение облигации";
+            case AMORTIZATION -> "Амортизация облигаци";
+            case TAX -> "Удержание налога";
+            default -> null;
+        };
+    }
+
+    private String getSecurityName(SecurityEventCashFlow cash) {
+       return securityRepository.findByIsin(cash.getIsin())
+                .map(SecurityEntity::getName)
+                .orElse(cash.getIsin());
     }
 }
